invisible-text-watermark/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invisible-text-watermark"
version = "0.1.0"
description = "Zero-width Unicode text watermarking with metadata payload, CRC integrity, and LLM SDK integrations."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
google = ["google-generativeai>=0.5"]
all = ["openai>=1.0", "anthropic>=0.20", "google-generativeai>=0.5"]
dev = ["pytest>=8.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/invisible_text_watermark"]
